fixed anomaly controller and .....

+++ services/anomaly.service.js
@@ -6,15 +6,39 @@ class AnomaliesService {
     }
 
     async getById(id) {
+        const anomaly = await Anomaly.findById(id);
 
-        return await Anomaly.findById(id);
+        if (!anomaly) {
+            const error = new Error('ANOMALY_NOT_FOUND');
+            error.code = 'ANOMALY_NOT_FOUND';
+            throw error;
+        }
+
+        return anomaly;
     }
 
     async create(data) {
+
+        const existing = await Anomaly.findOne({ name: data.name });
+
+        if (existing) {
+            const error = new Error('ANOMALY_ALREADY_EXISTS');
+            error.code = 'ANOMALY_ALREADY_EXISTS';
+            throw error;
+        }
+
         return await Anomaly.create(data);
     }
 
     async update(id, data) {
+
+        const existing = await Anomaly.findById(id);
+
+        if (!existing) {
+            const error = new Error('ANOMALY_NOT_FOUND');
+            error.code = 'ANOMALY_NOT_FOUND';
+            throw error;
+        }
 
         return await Anomaly.findByIdAndUpdate(
             id,
@@ -26,6 +50,13 @@ class AnomaliesService {
     }
 
     async delete(id) {
+        const existing = await Anomaly.findById(id);
+
+        if (!existing) {
+            const error = new Error('ANOMALY_NOT_FOUND');
+            error.code = 'ANOMALY_NOT_FOUND';
+            throw error;
+        }
 
         return await Anomaly.findByIdAndDelete(id);
     }
